Use same filtering logic for both convert and upload forms
